refpolicy git: update SRCREV and PV to 2.20151208
Issue: LIN10-1946

Update git repos rev to latest stable release and use PV to indicate
the exact release version just like all of the *_git.bb recipes has
been done in oe-core.

ref: https://github.com/TresysTechnology/refpolicy/wiki

And remove fix-ptest-failures-when-selinux-enabled.patch since it is
already in the source.

(LOCAL REV: NOT UPSTREAM) -- Sent to yocto on 20160107

Signed-off-by: Wenzong Fan <wenzong.fan@windriver.com>