libsemanage: fix race issue in parallel build
Issue: LIN1019-2979
Issue: LINCD-467

The previous patch to fix this issue is not correct. Actually, the
install-pywarp target doesn't need to depend on swigify target because
the semanage.py is not generated by swigify target but pywrap target.

Here is the dependency chain:
  install-pywrap -> pywrap -> $(SWIGSO) -> $(SWIGLOBJ) -> $(SWIGCOUT)
  -> semanage.py

But in the recipe, the swigify target is added explicitly in do_install:
  do_install_append() {
    oe_runmake install-pywrap swigify \
    [snip]
  }

This target will regenerate the semanage.py when do_install. So there
will be a potential race issue in parallel build. The install-pywrap
target is trying to install semanage.py when swigify target is
generating the file. Then an empty semanage.py will be installed. Remove
the target swigify to fix this issue.

(LOCAL REV: NOT UPSTREAM) -- sent to yocto on 20191204

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>